Refactored Dispatchers - LbinBaseDispatcherApplication and LibApplicationDispatcher are merged with ComApplicationDispatcher - removed application from libraries/default
