+++ src/components/com_application/templates/filters/html.php
@@ -1,30 +1,15 @@
 <?php
 
 /**
- * LICENSE: ##LICENSE##.
- *
  * @category   Anahita
- *
- * @author     Arash Sanieyan <ash@anahitapolis.com>
  * @author     Rastin Mehr <rastin@anahitapolis.com>
+ * @author     Arash Sanieyan <ash@anahitapolis.com>
  * @copyright  2008 - 2010 rmdStudio Inc./Peerglobe Technology Inc
  * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
  *
- * @version    SVN: $Id$
- *
  * @link       http://www.GetAnahita.com
  */
-
-/**
- * @category   Anahita
- *
- * @author     Arash Sanieyan <ash@anahitapolis.com>
- * @author     Rastin Mehr <rastin@anahitapolis.com>
- * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
- *
- * @link       http://www.GetAnahita.com
- */
-class LibApplicationTemplateFilterHtml extends KTemplateFilterAbstract implements KTemplateFilterWrite
+class ComApplicationTemplateFilterHtml extends KTemplateFilterAbstract implements KTemplateFilterWrite
 {
     /**
      * Convert the alias.
