+++ src/components/com_application/templates/helpers/render.php
@@ -1,21 +1,6 @@
 <?php
 
 /**
- * LICENSE: ##LICENSE##.
- *
- * @category   Anahita
- *
- * @author     Arash Sanieyan <ash@anahitapolis.com>
- * @author     Rastin Mehr <rastin@anahitapolis.com>
- * @copyright  2008 - 2010 rmdStudio Inc./Peerglobe Technology Inc
- * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
- *
- * @version    SVN: $Id$
- *
- * @link       http://www.GetAnahita.com
- */
-
-/**
  * Rendering script.
  *
  * @category   Anahita
@@ -26,7 +11,7 @@
  *
  * @link       http://www.GetAnahita.com
  */
-class LibApplicationTemplateHelperRender extends KTemplateHelperAbstract
+class ComApplicationTemplateHelperRender extends KTemplateHelperAbstract
 {
     /**
      * Template parameters.
