+++ src/components/com_application/views/json.php
@@ -1,20 +1,5 @@
 <?php
 
-/** 
- * LICENSE: ##LICENSE##.
- * 
- * @category   Anahita
- *
- * @author     Arash Sanieyan <ash@anahitapolis.com>
- * @author     Rastin Mehr <rastin@anahitapolis.com>
- * @copyright  2008 - 2010 rmdStudio Inc./Peerglobe Technology Inc
- * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
- *
- * @version    SVN: $Id: view.php 13650 2012-04-11 08:56:41Z asanieyan $
- *
- * @link       http://www.GetAnahita.com
- */
-
 /**
  * JSON view.
  *
@@ -26,18 +11,18 @@
  *
  * @link       http://www.GetAnahita.com
  */
-class LibApplicationViewJson extends LibBaseViewJson
+class ComApplicationViewJson extends LibBaseViewJson
 {
     /**
      * Content.
-     * 
+     *
      * var string|KException
      */
     public $content;
 
     /**
      * Just return the content.
-     * 
+     *
      * @return string
      */
     public function display()
@@ -45,7 +30,7 @@ class LibApplicationViewJson extends LibBaseViewJson
         $this->output = $this->content;
 
         //if content is an exception then
-        //handle               
+        //handle
         if ($this->content instanceof Exception) {
             $data = array();
             $data['code'] = $this->content->getCode();
@@ -62,9 +47,9 @@ class LibApplicationViewJson extends LibBaseViewJson
 
     /**
      * Converts an array of error into an array of serializable data.
-     * 
+     *
      * @param array $error An array of errors
-     * 
+     *
      * @return array
      */
     protected function _toData($errors)
