+++ src/templates/shiraz/params.ini
@@ -1,5 +1,5 @@
 cssStyle=style1
-compilestyle=1
+compilestyle=0
 compresstyle=1
 favicon=
 navbarInverse=0
